requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
flask-limiter==3.9.0
gunicorn==23.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
PyPDF2==3.0.1
Pillow==11.1.0
PyJWT==2.9.0
werkzeug>=3.0.0